refactor: extract dns-over-https types into dns namespace
